apps/simulator/pyproject.toml
[project]
name = "pulse-simulator"
version = "0.1.0"
description = "PULSE simulator: synthesises IoT, CCTV, and fan signals on a scripted match-day timeline."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "google-cloud-pubsub>=2.26.0",
  "google-cloud-firestore>=2.19.0",
  "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
